Major changes
Thought it was time for some major changes. I removed the "home" button
as it was redundant most of the time. Only for the single page or single
post view the button was kept, but moved in place of the "menu" button.
The author link was removed from the postlist, instead we now have
different post types. For example "post" (normal blog post), "link"
(quick links with short description), "post-it" (short post, twitter
like) and "quote". These require a certain structure in the "link" and
"post-it" quotes (see Readme). Also cleaned up the HTML and removed
unused CSS.